requirements_win.txt
cachetools==4.2.4
future==0.18.2
lxml==4.8.0
pdf2image==1.16.0
pefile==2021.9.3
Pillow==8.4.0
py2exe==0.11.0.1
python-docx==0.8.11
pywin32==303
